+++ src/main/java/ph/devcon/rapidpass/enums/RegistrarUserSource.java
@@ -1,0 +1,12 @@
+package ph.devcon.rapidpass.enums;
+
+public enum RegistrarUserSource {
+    /**
+     * Means that the registrar user was registered via bulk registration.
+     */
+    BULK,
+    /**
+     * Means that the registrar user was registered manually, through an online registration process.
+     */
+    ONLINE
+}