+++ src/main/java/ph/devcon/rapidpass/services/RegistryService.java
@@ -1,6 +1,7 @@
 package ph.devcon.rapidpass.services;
 
 import com.google.zxing.WriterException;
+import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -10,13 +11,11 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.repository.support.PageableExecutionUtils;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import ph.devcon.rapidpass.entities.AccessPass;
-import ph.devcon.rapidpass.entities.ControlCode;
-import ph.devcon.rapidpass.entities.Registrant;
-import ph.devcon.rapidpass.entities.ScannerDevice;
+import ph.devcon.rapidpass.entities.*;
 import ph.devcon.rapidpass.enums.AccessPassStatus;
 import ph.devcon.rapidpass.enums.PassType;
 import ph.devcon.rapidpass.enums.RecordSource;
+import ph.devcon.rapidpass.enums.RegistrarUserSource;
 import ph.devcon.rapidpass.kafka.RapidPassEventProducer;
 import ph.devcon.rapidpass.kafka.RapidPassRequestProducer;
 import ph.devcon.rapidpass.models.*;
@@ -26,6 +25,7 @@ import ph.devcon.rapidpass.utilities.StringFormatter;
 import ph.devcon.rapidpass.validators.StandardDataBindingValidation;
 import ph.devcon.rapidpass.validators.entities.BatchAccessPassRequestValidator;
 import ph.devcon.rapidpass.validators.entities.NewSingleAccessPassRequestValidator;
+import ph.devcon.rapidpass.validators.entities.agencyuser.BatchAgencyUserRequestValidator;
 
 import javax.validation.constraints.NotEmpty;
 import java.io.IOException;
@@ -58,13 +58,18 @@ public class RegistryService {
     private final RapidPassRequestProducer requestProducer;
     private final RapidPassEventProducer eventProducer;
 
+    private final AccessPassEventRepository accessPassEventRepository;
+    private final ApproverAuthService authService;
+    private final LookupTableService lookupTableService;
+    private final AccessPassNotifierService accessPassNotifierService;
+    private final RegistrarRepository registrarRepository;
+
     private final RegistryRepository registryRepository;
     private final ControlCodeService controlCodeService;
     private final RegistrantRepository registrantRepository;
-    private final LookupTableService lookupTableService;
     private final AccessPassRepository accessPassRepository;
-    private final AccessPassNotifierService accessPassNotifierService;
     private final ScannerDeviceRepository scannerDeviceRepository;
+    private final RegistrarUserRepository registrarUserRepository;
 
     /**
      * Creates a new {@link RapidPass} with PENDING status.
@@ -529,7 +534,7 @@ public class RegistryService {
      * @param approvedRapidPasses Iterable<RapidPass> of Approved passes application
      * @return a list of generated rapid passes, whose status are all approved.
      */
-    public List<String> batchUpload(List<RapidPassCSVdata> approvedRapidPasses) throws RegistryService.UpdateAccessPassException {
+    public List<String> batchUploadRapidPassRequest(List<RapidPassCSVdata> approvedRapidPasses) throws RegistryService.UpdateAccessPassException {
 
         log.info("Process Batch Approving of AccessPass");
         List<String> passes = new ArrayList<>();
@@ -606,6 +611,34 @@ public class RegistryService {
         return passes;
     }
 
+    public List<String> batchUploadApprovers(List<AgencyUser> agencyUsers) {
+        log.info("Processing batch registration of approvers.");
+        List<String> result = new ArrayList<String>();
+
+        // Validation
+        BatchAgencyUserRequestValidator newAccessPassRequestValidator = new BatchAgencyUserRequestValidator(this.registrarUserRepository, this.registrarRepository);
+
+        int counter = 1;
+        for (AgencyUser agencyUser : agencyUsers) {
+            try {
+
+                agencyUser.setUsername(StringUtils.trim(agencyUser.getUsername()));
+                agencyUser.setFirstName(StringUtils.trim(agencyUser.getFirstName()));
+                agencyUser.setLastName(StringUtils.trim(agencyUser.getLastName()));
+
+                agencyUser.setSource(RegistrarUserSource.BULK.name());
+                StandardDataBindingValidation validation = new StandardDataBindingValidation(newAccessPassRequestValidator);
+                validation.validate(agencyUser);
+
+                RegistrarUser registrarUser = this.authService.createAgencyCredentials(agencyUser);
+
+                result.add("Record " + counter++ + ": Success. ");
+            } catch ( Exception e ) {
+                result.add("Record " + counter++ + ": Failed. " + e.getMessage());
+            }
+        }
+        return result;
+    }
 
     /**
      * This is thrown when updates are not allowed for the AccessPass.
@@ -647,5 +680,14 @@ public class RegistryService {
         device.setStatus(request.getStatus());
 
         return scannerDeviceRepository.saveAndFlush(device);
+    }
+
+    public RapidPassEventLog getAccessPassEvent(Integer eventId, Pageable pageable) {
+        Page<AccessPassEvent> accessPassEvents = accessPassEventRepository.findAllByIdIsGreaterThanEqual(eventId, pageable);
+        if (accessPassEvents == null || accessPassEvents.isEmpty()) {
+            return null;
+        } else {
+            return RapidPassEventLog.buildFrom(accessPassEvents, controlCodeService);
+        }
     }
 }