+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2020-04-12
+
+## [1.3.1-SNAPSHOT]
+### Added
+- [#172](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/172) - Can now bulk register approvers. See 
+    `src/main/resources/approver-bulk-registration.csv` for template CSV.
+
 ## [1.3.0-SNAPSHOT]
 ### Changed
 - Bulk upload validation has been optimized to not query database if basic validation fails.
@@ -14,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - introduced 'bulk-upload.process' parameter to switch between kafka-based bulk upload or internal processing.
 - added index on access_pass to optimize queries. 
 - [#345](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/345) fixed `get /registry/access-passes` to allow multiple apor types in the query parameters
+- [#350](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/350) - added new endpoint for syncing checkpoint devices 'get /batch/access-pass-events'
 
 ## [1.2.1] - 2020-04-11
 ### Fixed